consolekit: Remove bbappend as gnome-common is not GPLv3
gnome-common's LICENSE fixed with this oe-core commit:
commit 97469db0cba6d21ab1290238225d3fe365931a5c
Author: Ross Burton <ross.burton@intel.com>
Date:   Tue Oct 30 12:02:45 2012 +0000

gnome-common: Fix license
gnome-common 2.28 is GPLv2+.  From Christian Persch, upstream:
The licence is presumed GPL2+, although it's not there explicitly. GPL2+
because as far as I could figure out when I tried to, gnome-autogen started in
gnome-core which had a GPL2 COPYING file.

Signed-off-by: Andrei Gherzan <andrei.gherzan@windriver.com>