packagegroups: restructure the yocto-ivi PGs
with the current implementation of packagegroups, it's not clear how a
particular recipe map to the GENIVI baseline document.

this scheme remove confusion about which packagegroup a certain recipe belong
to. And provide guidance to the customers about modifying the image. For
instance customers should be careful about replacing specific-p1 components if
they wish to maintain GENIVI compliance.

Signed-off-by: Fahad Usman <fahad_usman@mentor.com>