ecryptfs-utils: Modify systemd service file to 'simple'.
ecryptfsd is a resident program, i.e. daemon. According to systemd
documentation, Type=oneshot/RemainAfterExit=yes should be used for programs
that do not remain working after exit. But for daemons we should use
Type=forked to indicate that the service remains running in background.
ecryptfs provides option -f to run in foreground. This option is preferred
when the service is started from init/systemd. That is why we use
Type=simple (default) to indicate that the service runs in foreground. To
manage the service properly systemd uses its type. E.g. if ecryptfsd
silently dies, systemd will recognize its failure in case of simple or
forked, but not in case of oneshot.

Signed-off-by: Mikhail Durnev <mikhail_durnev@mentor.com>
Signed-off-by: Noor Ahsan <noor_ahsan@mentor.com>