common-api-c++: added CommonAPI runtime
This is the CommonAPI base C++ library, which provides the application
interface for users and can load runtime bindings such as dbus.

Common API and its mechanism specific bindings (e.g. Common API D-Bus)
provide a set of libraries and tools to work with RPC communication in
a way independent of wich mechanism is used.

References:
- http://projects.genivi.org/commonapi/
- http://lists.genivi.org/pipermail/genivi-ipc/2013-March/000057.html

Signed-off-by: Holger Behrens <Holger.Behrens@windriver.com>