common-api-c++-dbus: added CommonAPI D-Bus binding
This is the CommonAPI D-Bus binding C++ library, which provides the necessary
code to communicate over D-Bus. This is invisible to the application code,
and simply needs to be linked against.

Common API and its mechanism specific bindings (e.g. Common API D-Bus)
provide a set of libraries and tools to work with RPC communication in
a way independent of wich mechanism is used.

The D-Bus marshaling patch adds functions to support querying and manipulating
the message body and signature. This is useful for code generators, which can
generate custom marshaling functions based on a given IDL. Those functions tend
to be optimized and faster than the generic iterator based marshaling.

References:
- http://projects.genivi.org/commonapi/
- http://lists.genivi.org/pipermail/genivi-ipc/2013-March/000057.html

Signed-off-by: Holger Behrens <Holger.Behrens@windriver.com>