Use weston version 1.4% from meta-ivi
PREFERRED_VERSION variables are only found in .conf files.
Removing this variable occurence from the .bb image file and
defining it in the conf/layer.conf file. Even though at the moment
it is not absolutely necessary to define it there because meta-ivi
has a higher BBFILE_PRIORITY than that of poky, we define it in the
conf file to make the statement more clear to the users of the layer.

Signed-off-by: Florin Sarbu <florin.sarbu@windriver.com>
Acked-by: Andrei Gherzan <andrei.gherzan@windriver.com>