weston-ivi-shell: Add patches on top of weston v1.5
Instead of using another upstream repo, we use the one from yocto and only add
the needed patches for weston-ivi-shell.

Repo: https://github.com/Tarnyko/weston-ivi-shell
Branch: weston150_ivi-shell_v6

One of the patches can't be applied as quilt, the default yocto patch tool,
doesn't support binary diffs. The solution would be to use git as patch tool
but this machanism is broken due to a yocto bug:
https://bugzilla.yoctoproject.org/show_bug.cgi?id=5968
This patch includes a workaround for this situation: manually copy png files
from 0007-data-add-reference-image-files-for-weston-ivi-shell-.patch .

Signed-off-by: Andrei Gherzan <andrei.gherzan@windriver.com>