qtwayland: Do not compile xcomposite-egl client buffer integration.
In our platform we use the wayland-egl client buffer integration.
Because at this point we do not use the xcomposite-egl client buffer integration,
and because it fails to compile with the error:

src/hardwareintegration/client/xcomposite-egl/qwaylandxcompositeeglclientbufferintegration.cpp:137:67: error: invalid conversion from 'Display* {aka _XDisplay*}' to 'EGLNativeDisplayType {aka int}' [-fpermissive]
|      integration->mEglDisplay = eglGetDisplay(integration->mDisplay);

we skip past it and do not compile it.

Signed-off-by: Florin Sarbu <florin.sarbu@windriver.com>
Signed-off-by: Holger Behrens <holger.behrens@windriver.com>