Hacks for weston and browser-poc
Added one browser-poc patch that diminishes the size of the bookmarks window.
The qml window for bookmarks is contained into a 0, 0, 1024, 80 rectangle and for
now we cannot dynamically change it's size. For this reason, we resize the bookmarks window.

Added one weston patch that modifies the tiling of windows so that we can properly show the
browser-poc which is comprised of 2 windows: one for address, back and forward button, bookmarks
and the other window which does the rendering.

Signed-off-by: Florin Sarbu <florin.sarbu@windriver.com>
Signed-off-by: Holger Behrens <holger.behrens@windriver.com>