persistence-administrator: upgrade to 1.0.8
"
Subject: [meta-ivi]  Persistence components in meta-ivi
From: Igor Socec <igor.socec@pelagicore.com>

I'm working on a project that uses meta-ivi 10.0 and Persistence
components. There are some critical bugfixes in Persistence
Administrator which are not available in meta-ivi. There is also a
bugfix in Persistence Common Object available upstream that's missing
in meta-ivi.

...<skip>

Here is a stack of upstream commits that we apply on top of meta-ivi
10.0 to resolve bugs in Persistence:

persistence-administrator:
- add-missing-compiler-flags-for-CPP-compliance.patch
- PAS-Vers.-1.0.6-Fixed-bug-295-further-calls-of-persa.patch
- PAS-Vers.-1.0.7-Fixed-bug-299-Link-to-shared-group-d.patch
- Takeover-fix-for-Bug-299-from-Continental-git-branch.patch
- PAS-Vers.-1.0.8-Fixed-bug-437-pas-daemon-stops-trigg.patch

persistence-common-object:
- fixed GENIVI Bug 345
"

Signed-off-by: Yong-iL Joh <yong-il.joh@windriver.com>