checkout script: remove bash specific syntax
Script used >& syntax but did not specify a shell that is guaranteed
to understand it (csh, bash, zsh).  This triggers the message
"Syntax error: Bad fd number" on some build machines.

An alternative would be to use /bin/bash explicitly since ci-build
requires bash anyway, but maybe checkout might be used on its own.
This is just as easy.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>