ci-build: Fix the git diff reporting on conf files
There is a git diff performed and stored in a text file.  This is to
note in build results if there were any local conf changes.  It was
wrong (based on GDP directory structure) and also syntactically wrong,
which caused the command to fail, and through the -e flag consequently a
script error exit as well.

The script also needs to exit with success if reaching the end (cleanup
function might report some other return code)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>