README.md: Final tweaks on branch and tag policies
Reorder the branch policy statements in (I think) a more logical order
- first principles (cherry-pick), then versioning, etc.  Minor word
changes to make the text shorter and clearer.  Add note about Yocto/Poky
upgrades, as introduction to the next point which talks about
-rocko/-sumo branch names.  Remove vexpressa9 from list of tested
platforms.  Add a note on how other hardware is supported.

In other sections, minor word changes and whitespace.

(This is a backport of b181b5fdde from master)

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>