checkout_layer_hash.sh: Fetch selected branch
The CI system will typically make shallow clones, and it might be set to
fetch a specific layer branch.  In a transition period, when the README
specifies a new branch we might lack the requested commit hash, but
since the README lists the branch name, we can use it and make sure it
is fetched first.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>