dbus.bbappend: Remove patch for readdir/readdir_r
A patch was applied previously to make dbus 1.10 compile correctly in
the presence of the readdir_r call (and only for qemux86, 32-bit).

The update to poky/sumo now builds dbus 1.12 Between 1.10 and 1.12 the
code has moved and more importantly the readdir_r seems to have been
changed back to readdir, so the patch cannot apply.  Let's try removing
the patch.

The single patch however mixed different things.  Testing shows that
the signed/unsigned mismatch is still a problem (warning which is
treated as an error) but only on 32-bit compilation. This part of the
patch is therefore kept for now.   I have not deeply analyzed the effect
of this cast, just keeping what was already there.

[BASE-63] master fails application of dbus patch for qemux86 target

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>