test-image: Set constant PV to avoid basehash errors
Yocto does not allow the values to change during the build - it causes
the "basehash" type error if they do.  PV for image definition is not an
important value, in fact I have been advivsed that it could be omitted
since the recipe does not deal with a single "package".

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>