ci-build: Add meta-ivi-test layer (conditionally)
If it's test-image build then append the test layer to BBLAYERS
variable.  It cannot be permanently added to bblayers.conf.sample since
some recipes modify the included components.

Added a function for append_bblayers_conf analogous to append_local_conf
for this purpose.  The rest of the modified lines are cosmetic
improvements.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>