persistence-administrator: version 1.0.10
(A.k.a. Persistence Administration Service (PAS), but the bb component
 name is persistence-administrator for some historical reason)

It was updated from version 1.0.9, but for some reason the repository is
not tagged with that version, although the commmit message itself
specifies the version.  There is now a tag created for 1.0.10

Short log of changes - the gdbus-codegen update is the important one

ebfb5dd README: fix headline and minor cleanups
49b7936 Makefile: fix gdbus-codegen invocation

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>