dlt-daemon: bump to 2.18.1
Changed SRC_URI to release tarballs as it is faster and more intuitive.

Insallation of dlt-dbus.service was removed since package is built
with WITH_DLT_DBUS=OFF by default now (see @1b3c8b3d10fbc).

Removed mainstreamed 0001-Fix-compilation-with-glibc-2.28.patch, and
rebased 0001-change-file-position-for-dlt-test-filetransfer.patch
from meta-ivi-test.

Also modified dlt-daemon_t.inc to test success or failure of
dlt-test-user to what is done in .travis.xml in dlt-daemon repository.
Output has changed so that more lines outputted that are not filtered
which means the test fails without this change.

Co-authored-by: Martin Ejdestig <mejdestig@luxoft.com>
Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>