dlt-daemon: bump recipe content to 2.18.4
The SRC_URI checks are update to match the 2.18.4 version.

Since 2.18.3, the installation of systemd services was removed when
the corresponding package isn't built (e.g dlt-system.service isn't
installed if the dlt-daemon isn't compiled with WITH_DLT_SYSTEM=ON).
See @514717f2.
Therefore, I use the PACKAGECONFIG facility of yocto to choose which
service file should be installed based on which feature has been
activated by the user.

As I use the PACKAGECONFIG facility, I also populate the CMake flags
based on which feature is activated or not.

The dlt-daemon-systemd package is removed from the packagegroup. The
packagegroup can't be aware that, sometimes, the package can be empty
(depending of which features are activated or not). In return, it's the
dlt-daemon package itself which decides when the dlt-daemon-systemd
package should be added in the final image or not

Signed-off-by: Sebastien RAILLET <sebastien.raillet@marelli.com>